requirements.txt
PyMuPDF==1.18.19
faiss-cpu==1.7.1
pytest
autoflake
isort
black
streamlit
python-dotenv